Endline detection tune
Endline detection was getting tripped on the outline of the red light being emitted from traffic light. I upped the minimmum contour area that needs to be seen for endline to be detected, since I originally calibrated it for 2 inch thick magenta lines, when in fact its 6 inches thick, so we can afford it make it more aggressive.

I also made it so endline, lane detection, and sign detection will only be instatiated after traffic light is detected. This removes the chance that endline detection trips on the traffic light and also makes sense to do imo